backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
Authlib==1.6.0
bcrypt==3.2.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
confluent-kafka==2.11.0
cryptography==45.0.4
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.13
fastapi-cli==0.0.7
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
greenlet==3.2.3
gTTS==2.5.4
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
moviepy==2.2.1
multidict==6.5.0
numpy==2.3.1
openai==1.91.0
orjson==3.10.18
packaging==25.0
passlib==1.7.4
pillow==11.2.1
proglog==0.1.12
propcache==0.3.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.0
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
stripe==8.9.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
